feature: 完成了北向保修事务管理的API

+++ server_event_manager/event_manage.py
@@ -15,11 +15,13 @@ async def init_table(app) -> None:
             CREATE TABLE IF NOT EXISTS event (
                 id_event BIGINT PRIMARY KEY,
                 id_room VARCHAR(20) NOT NULL,
+                id_initiator VARCHAR(20) NOT NULL,
                 time_start VARCHAR(25) NOT NULL,
                 time_end VARCHAR(25) DEFAULT '',
                 event VARCHAR(255) NOT NULL,
                 result VARCHAR(255) DEFAULT '',
-                state VARCHAR(10) NOT NULL
+                state VARCHAR(10) NOT NULL,
+                id_processor VARCHAR(20) DEFAULT ''
             );
         """
         await app.ctx.db.execute(sql)
@@ -32,23 +34,23 @@ def gen_event_id() -> str:
 
 class event_manager_view(HTTPMethodView):
 
-    async def post(self, request, id) -> JSONResponse:
+    async def post(self, request) -> JSONResponse:
         req = request.json
         sql = """
-            SELECT COUNT(*) FROM event WHERE id_room = %s AND event = %s;
+            SELECT COUNT(*) FROM event WHERE id_room = %s AND event = %s AND state = 'pending';
         """
         try:
-             result = await request.app.ctx.db.fetch(sql, (id, req["event"]))
+             result = await request.app.ctx.db.fetch(sql, (req["id_room"], req["event"]))
         except Exception as e:
             return json({"msg": "Event add fail, error: {}".format(str(e))}, status=400)
         if result[0]["COUNT(*)"] > 0:
             return json({"msg": "Do not add events repeatedly!"}, status=400)
         sql = """
-            INSERT INTO event (id_event, id_room, time_start, event, state)
-            VALUES (%s, %s, %s, %s, %s);
+            INSERT INTO event (id_event, id_room, id_initiator, time_start, event, state)
+            VALUES (%s, %s, %s, %s, %s, %s);
         """
         try:
-            await request.app.ctx.db.execute(sql, (gen_event_id(), id, req["time_start"], req["event"], req["state"]))
+            await request.app.ctx.db.execute(sql, (gen_event_id(), req["id_room"], req["id_initiator"], req["time_start"], req["event"], req["state"]))
         except Exception as e:
             return json({"msg": "Event add fail, error: {}".format(str(e))}, status=400)
         return json({"msg": "Event added successfully"})
@@ -117,17 +119,17 @@ class event_manager_view(HTTPMethodView):
         if result[0]["COUNT(*)"] == 0:
             return json({"msg": "Event not exist!"}, status=410)
         sql = """
-            UPDATE event SET time_end = %s, result = %s, state = %s WHERE id_event = %s;
+            UPDATE event SET time_end = %s, result = %s, state = %s, id_processor = %s WHERE id_event = %s;
         """
         try:
-            await request.app.ctx.db.execute(sql, (req["time_end"], req["result"], req["state"], id))
+            await request.app.ctx.db.execute(sql, (req["time_end"], req["result"], req["state"], req["id_processor"], id))
         except Exception as e:
             return json({"msg": "Event update fail, error: {}".format(str(e))}, status=400)
         return json({"msg": "Event update successfully"})
     
     async def delete(self, request, id) -> JSONResponse:
         sql = """
-            SELECT * FROM event WHERE id_event = %s;
+            SELECT state FROM event WHERE id_event = %s;
         """
         id = int(id)
         try:
@@ -136,6 +138,8 @@ class event_manager_view(HTTPMethodView):
             return json({"msg": "Event delete fail, error: {}".format(str(e))}, status=400)
         if len(result) == 0:
             return json({"msg": "Event not exist!"}, status=410)
+        if result[0]["state"] == "processed":
+            return json({"msg": "Processed events cannot be deleted!"}, status=405)
         sql = """
             DELETE FROM event WHERE id_event = %s;
         """
